添加jwt token 认证

+++ src/main/java/com/example/security/JwtUser.java
@@ -1,5 +1,6 @@
 package com.example.security;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.Date;
 
@@ -10,7 +11,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 
 public class JwtUser implements UserDetails {
 
-    private final Long id;
+    private final BigDecimal id;
     private final String username;
     private final String firstname;
     private final String lastname;
@@ -21,7 +22,7 @@ public class JwtUser implements UserDetails {
     private final Date lastPasswordResetDate;
 
     public JwtUser(
-            Long id,
+            BigDecimal id,
             String username,
             String firstname,
             String lastname,
@@ -42,7 +43,7 @@ public class JwtUser implements UserDetails {
     }
 
     @JsonIgnore
-    public Long getId() {
+    public BigDecimal getId() {
         return id;
     }
